requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
email-validator>=2.1.0
aiosmtplib>=3.0.0

